+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - CLI command to emulate stop button being pressed.
 - CLI command to print active alarms.
+- CLI now also sends Serial_change_character when alarm configuration changes.
 
 
 ## [0.6.1] - 2022-04-27